enregistrement commande dans bdd

+++ LaraShop6/app/Http/Controllers/CheckoutController.php
@@ -1,12 +1,14 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use DateTime;
+use App\Order;
 use Stripe\Stripe;
 use Stripe\PaymentIntent;
+use Illuminate\Support\Arr;
 use Illuminate\Http\Request;
 use Gloudemans\Shoppingcart\Facades\Cart;
-use Illuminate\Support\Arr;
+use Illuminate\Support\Facades\Session;
 
 class CheckoutController extends Controller
 {
@@ -54,9 +56,42 @@ class CheckoutController extends Controller
      */
     public function store(Request $request)
     {   
-        Cart::destroy();
         $data = $request->json()->all();
-        return $data['paymentIntent'];
+        $order = new Order();
+        $order->payment_intent_id = $data['paymentIntent']['id'];
+        $order->amount = $data['paymentIntent']['amount'];
+        $order->payment_created_at = (new DateTime())
+            ->setTimestamp($data['paymentIntent']['created'])
+            ->format('Y-m-d H:i:s');
+
+
+        $products = [];
+        $i = 0;
+        foreach(Cart::content() as $product) {
+            $products['product_' . $i][] = $product->model->title;
+            $products['product_' . $i][] = $product->model->price;
+            $products['product_' . $i][] = $product->qty;
+            $i++;
+        }
+        
+        $order->products = serialize($products);
+        $order->user_id = 15;
+        $order->save();
+        
+
+        if($data['paymentIntent']['status'] === 'succeeded') {
+            Cart::destroy();
+            Session::flash('success', 'Votre commande à été traitée avec succés');
+            return response()->json(['success' => 'Payment Intent Succeeded']);
+        } else {
+            return response()->json(['error' => 'Payment Intent Failed']);
+        }
+        
+    }
+
+
+    public function thankyou() {
+        return Session::has('success') ? view('checkout.thankyou') : redirect()->route('shop');
     }
 
     /**
